requirements.txt
astroid==2.3.3
certifi==2019.9.11
cheroot==7.0.0
Click==7.0
colorama==0.4.1
Django==2.2.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
isort==4.3.21
jaraco.functools==2.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
pipenv==2018.11.26
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
PyQt5==5.13.0
PyQt5-sip==4.19.19
pyqt5-tools==5.13.0.1.5
python-dotenv==0.10.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
web.py==0.40
websockets==8.0.2
wrapt==1.11.2
